Results somewhat making sense now... with caveats
Displacing atoms along polarization direction no red-shifts resonance, as expected, and displacing orthogonal to polarization direction does not, also as expected.  However, the dipole expectation value seems to go off the rails if the particles are closer together than ~70 atomic units, which is like 3.5 nm, so it just bears thinking if coupling strength is correct or not... actually seems fairly reasonable, but it would be good to confirm.