Revised FSSH code to compute forces using H and D in local basis
Now code initializes system in an eignenstate of the polariton Hamiltonian, which is a superposition in the local basis ( |g0>, |g1>, |e0>, and |e1> ).  The derivative coupling is also computed in the local basis.  The change in populations to determine hopping probability is tracked in the polariton basis, since this the basis in which the active state is an eigenstate.