polaritonic class re-write comple and ready for testing
Re-write assumes that the Hamiltonian matrix can be
non-Hermitian due to the fact that the photonic finite lifetime can be
represented as a complex frequency:
    omega = omega_r - i * gamma
where omega_r is the ordinary "real" frequency and gamma is related to the lifetime,
i.e. it is the same gamma we feed to the Lindblad operator.
The polariton basis is defined by diagonalizing this non-Hermitian total Hamiltonian
and the derivative couplings are computed using this basis.

To compute the Liouville-Lindblad equation of motion, the commutator
[H_total, D] only includes the REAL part of the H_total.

The polaritonic potential energy surfaces are computed as the absolute magnitude
of the complex energies that result from diagonalizing the non-Hermitian H_total

The nuclear forces arise from applying the Hellman-Feynman theorem in the local basis:
    F = -Tr(H'_local D^act_local)
    where H'_local = (H_local(R + dr) - H_local(R - dr))/2dr
and since the H_local(R + dr) and the H_local(R - dr) both have identical
non-Hermtian parts, they cancel and the Force will be real!

FURTHER NOTE: polaritonic_bup.py contains the original versions of functions
that neglected the complex frequency; one can refer to it if we decide this
non-Hermitian business is bad!